ISAICP-5377: Remove only files asked.

+++ web/modules/custom/joinup_core/joinup_core.post_update.php
@@ -683,18 +683,19 @@ function joinup_core_post_update_enable_joinup_privacy() {
 }
 
 /**
- * Deletes unused files.
+ * Deletes unused files explicitly requested for deletion.
  */
 function joinup_core_post_update_delete_orphaned_files() {
-  $query = \Drupal::database()->select('file_managed', 'fm')
-    ->fields('fm', ['fid']);
-  $query->leftJoin('file_usage', 'fu', 'fm.fid = fu.fid');
-  $query->leftJoin('node_revision', 'nr', 'fu.id = nr.nid');
-  $query->condition('fu.type', 'node');
-  $query->isNull('nr.nid');
-  $results = $query->execute()->fetchCol();
+  $files_to_remove = [
+    'public://document/2017-05/e-trustex_software_architecture_document_0.pdf',
+    'public://document/2013-12/e-TrustEx Interface Control Document.pdf',
+  ];
 
-  foreach (File::loadMultiple($results) as $file) {
-    $file->delete();
+  $file_storage = \Drupal::entityTypeManager()->getStorage('file');
+  foreach ($files_to_remove as $uri) {
+    $files = $file_storage->loadByProperties(['uri' => $uri]);
+    if ($file = reset($files)) {
+      $file->delete();
+    }
   }
 }
